+++ src/internals.h
@@ -21,8 +21,8 @@ Rcpp::List cpp_lookup_sequenced_ancestor(Rcpp::List data, Rcpp::List param, size
 
 void lookup_sequenced_ancestor(Rcpp::IntegerVector alpha, Rcpp::IntegerVector kappa, 
 			       Rcpp::LogicalVector has_dna, size_t i, 
-			       int *out_alpha, 
-			       int *out_n_generations, 
+			       size_t *out_alpha, 
+			       size_t *out_n_generations, 
 			       bool *found_sequenced_ancestor
 			       );
